# be/src/main/resources/api.yaml
openapi: 3.0.0

info:
  title: Employee Service API
  description: API description.
  version: 1.0.0

#servers:
#  - url: 'http://localhost:8080/api/v1'

paths:
  /employee:
    get:
      tags: [ employee ]
      operationId: getEmployees
      description: 'Returns employees'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'dto.yaml/#/components/schemas/EmployeeResponseDto'
    post:
      tags: [ employee ]
      operationId: createEmployee
      description: 'Create employee'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'dto.yaml#/components/schemas/EmployeeRequestDto'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: 'dto.yaml#/components/schemas/EmployeeResponseDto'
  /employee/{employeeId}:
    put:
      tags: [ employee ]
      operationId: updateEmployee
      description: 'Update employee'
      parameters:
        - $ref: '#/components/parameters/employeeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'dto.yaml#/components/schemas/EmployeeRequestUpdateDto'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: 'dto.yaml#/components/schemas/EmployeeResponseUpdateDto'
    delete:
      tags: [ employee ]
      operationId: deleteEmployee
      description: 'Delete employee'
      parameters:
        - $ref: '#/components/parameters/employeeIdParam'
      responses:
        '204':
          description: Deleted

components:
  parameters:
    employeeIdParam:
      name: employeeId
      in: path
      description: 'Employee id'
      required: true
      schema:
        type: string